Fixes for kotlin to java conversion
