+++ lib/src/main/kotlin/com/swmansion/starknet/data/Constants.kt
@@ -4,5 +4,10 @@ package com.swmansion.starknet.data
 
 import com.swmansion.starknet.data.types.Felt
 
+/**
+ * Declare sender address.
+ *
+ * Constant value to be set as sender address in declare transaction.
+ */
 @JvmField
 val DECLARE_SENDER_ADDRESS = Felt(1)
