+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Execution.kt
@@ -1,3 +1,4 @@
+@file:JvmName("Execution")
 package com.swmansion.starknet.data.types
 
 import com.swmansion.starknet.data.selectorFromName
