+++ lib/src/main/kotlin/com/swmansion/starknet/extensions/Base64Gzipped.kt
@@ -1,3 +1,4 @@
+@file:JvmName("Base64Gzipped")
 package com.swmansion.starknet.extensions
 
 import java.io.ByteArrayOutputStream
