+++ lib/src/main/kotlin/com/swmansion/starknet/extensions/Json.kt
@@ -1,3 +1,4 @@
+@file:JvmName("Json")
 package com.swmansion.starknet.extensions
 
 import com.swmansion.starknet.data.types.Felt
