+++ lib/src/main/kotlin/com/swmansion/starknet/provider/rpc/JsonRpcResponse.kt
@@ -2,6 +2,7 @@ package com.swmansion.starknet.provider.rpc
 
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.KSerializer
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
@@ -11,9 +12,16 @@ import kotlinx.serialization.json.JsonElement
 
 @Serializable
 data class JsonRpcResponse<T>(
+    @SerialName("id")
     val id: Int,
-    val jsonrpc: String,
+
+    @SerialName("jsonrpc")
+    val jsonRpc: String,
+
+    @SerialName("result")
     val result: T? = null,
+
+    @SerialName("error")
     val error: JsonElement? = null, // FIXME: Add error types
 )
